server/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
tensorflow==2.17.0
pillow==10.4.0
numpy==1.26.4
keras==3.6.0